Test with Java 21 (#127)
Java 21 was released Sep 19, 2023. We want to announce full support
for Java 21 in early October and would like the most used plugins to
be compiled and tested with Java 21.

The acceptance test harness and plugin bill of materials tests are
already passing with Java 21. This is a further step to improve plugin
readiness for use with Java 21 and for development with Java 21.